rename files following PINA standard
