+++ pina/problem/__init__.py
@@ -1,6 +1,7 @@
 """Module for the Problems."""
 
 __all__ = [
+    "ProblemInterface",
     "AbstractProblem",
     "SpatialProblem",
     "TimeDependentProblem",
@@ -8,6 +9,7 @@ __all__ = [
     "InverseProblem",
 ]
 
+from pina._src.problem.problem_interface import ProblemInterface
 from pina._src.problem.abstract_problem import AbstractProblem
 from pina._src.problem.spatial_problem import SpatialProblem
 from pina._src.problem.time_dependent_problem import TimeDependentProblem
